requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo
bson
pydantic
supabase
python-multipart
requests
